Move most RazorSyntaxNodeExtensions methods into the Workspaces layer
